requirements.txt
flask
supabase
pandas
numpy
scikit-learn
plotly
python-dotenv
openpyxl
joblib
